% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_tsv.R
\name{convertTSV}
\alias{convertTSV}
\title{convert TSV to CSV format}
\usage{
convertTSV(varDescr, datafile)
}
\arguments{
\item{varDescr}{variable description}

\item{datafile}{input TSV file path}
}
\value{
A list of status
}
\description{
Convert tab separated values to csv format
}
\examples{
 convertTSV([{"name":["indid"],"internalName":["indid"],"type":["numeric"],"dataType":[null]}],'D:\\\\mde\\F1.tsv')

}
