% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/destring.R
\name{destring}
\alias{destring}
\title{destring variables}
\usage{
destring(varDescr, datafile, updatedDataPath, listOfVariables)
}
\arguments{
\item{varDescr}{variable description}

\item{datafile}{input data(csv) file path}

\item{updatedDataPath}{Updated data file path}

\item{listOfVariables}{list of variables to perform destring operation}
}
\value{
A list contains the destringed variables
}
\description{
Convert the numeric variables to string variables
}
\examples{
 destring([{"name":["uqnr"],"internalName":["uqnr"],"type":["character"],"dataType":["character"]}],'D:\\\\mde\\\\F1(1).csv',['uqnr'])

}
