% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_data.R
\name{export}
\alias{export}
\title{export dataset to STATA/SPSS/CSV format}
\usage{
export(varDescr, datafile, type, versionNo, outputfile)
}
\arguments{
\item{varDescr}{variable description}

\item{datafile}{input data(csv) file path}

\item{type}{Type of data file (dta | sav | csv)}

\item{versionNo}{STATA version}

\item{outputfile}{Output file path}
}
\value{
A list contains the status
}
\description{
export dataset to STATA/SPSS/CSV format
}
\examples{
 export([{"name":["hhid"],"internalName":["hhid"],"labl":["Household identification"],"val":[],"dcml":["0"],"width":["16"],"type":["numeric"],"intrvl":["contin"],"dataType":["double"]}],'D:\\\\mde\\\\f1.csv', 'csv','D:\\\\out\\\\HND_2012_L2L.csv')

}
