% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_fwf.R
\name{exportFWF}
\alias{exportFWF}
\title{export dataset to fixed width text}
\usage{
exportFWF(varDescr, datafile, outpath)
}
\arguments{
\item{varDescr}{variable description}

\item{datafile}{input data(csv) file path}

\item{outpath}{output fwf file path}
}
\value{
A list of variable
}
\description{
export dataset to fixed width text format
}
\examples{
 exportFWF([{"name":["hhid"],"internalName":["hhid"],"catVal":[],"width":[4],"dcml":["0"],"StartPos":[1],"EndPos":[4],"type":["numeric"],"StorageType":"","ReadFormat":"","dataType":["double"]}],'D:\\\\mde\\\\f1.csv')

}
