% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{import}
\alias{import}
\title{Import dataset}
\usage{
import(filepath = "", type = "", fileId = "", freqLimit = "")
}
\arguments{
\item{filepath}{Path to data file}

\item{type}{Type of data file (dta | sav | csv)}

\item{fileId}{unique file id}

\item{freqLimit}{frequency limit}
}
\value{
A list with variable metadata
}
\description{
This package import datasets of different datatypes (STATA | SPSS | CSV)
}
\examples{
 import('D:\\\\MetadataEditorSource\\\\MetadataEditor_Dev\\\\MetadataEditor\\\\Sample Data\\\\in\\\\HND_2012_L2L.dta', 'dta', 'F1', 36000)

}
