% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resequence.R
\name{resequence}
\alias{resequence}
\title{resequence variables}
\usage{
resequence(varDescr, datafile)
}
\arguments{
\item{varDescr}{variable description}

\item{datafile}{input data(csv) file path}
}
\value{
A list of status
}
\description{
Calculate variable width, startpos endpos and update the variable info
}
\examples{
 resequence([{"name":["hhid"],"internalName":["hhid"],"dcml":["0"],"type":["numeric"],"catVal":[],"width":["16"],"StartPos":[null],"EndPos":[null],"dataType":[null]}],'D:\\\\F1.csv')

}
