% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/string_len.R
\name{stringLength}
\alias{stringLength}
\title{get string length}
\usage{
stringLength(listOfVariables, datafile)
}
\arguments{
\item{listOfVariables}{list of variables}

\item{datafile}{input data(csv) file path}
}
\value{
A list contains the max length of variables
}
\description{
Calculate the length of string variable
}
\examples{
 stringLength([ {"internalName" : "uqnr", "type" : "character"} ],'D:\\\\mde\\\\f1.csv')

}
