% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_stats.R
\name{updateStats}
\alias{updateStats}
\title{update variable statistics}
\usage{
updateStats(typeOfActions, csvpath)
}
\arguments{
\item{typeOfActions}{different type of actions}

\item{csvpath}{input csv file path}
}
\value{
A list of updated variable statistics
}
\description{
Update variable statistics when perform different actions
}
\examples{
 updateStats([{"action":"encodeMissing","name":"REGION","properties":[{"op":"=","val":"1"}]}],'D:\\\\mde\\\\F1.csv')

}
