% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{writeCSV}
\alias{writeCSV}
\title{Write data file(CSV)}
\usage{
writeCSV(filepath = "", type = "", csvPath = "")
}
\arguments{
\item{filepath}{Path to data file}

\item{type}{Type of data file (dta | sav | csv)}

\item{csvFile}{Output file path}
}
\value{
A list contains the status
}
\description{
This package import dataset and write data in csv format
}
\examples{
 writeCSV('D:\\\\MetadataEditorSource\\\\MetadataEditor_Dev\\\\MetadataEditor\\\\Sample Data\\\\in\\\\HND_2012_L2L.dta', 'dta','D:\\\\MetadataEditorSource\\\\MetadataEditor_Dev\\\\MetadataEditor\\\\Sample Data\\\\in\\\\HND_2012_L2L.csv')

}
